requirements.txt
PyDispatcher==2.0.6
Scrapy==2.8.0
folium~=0.14.0
ipinfo~=4.4.2
pytest~=7.2.1